Anchor the atomic job-plus-batch terminal test in the batch contract
markJobSucceededAndUpdateBatch needs a live batch to increment, so its
test set up and asserted everything through the optional batch view while
sitting in the core terminal contract. A store without the batch
capability skipped it, leaving the atomic job-side transition unverified.

Move the test to the batch contract, where the batch view is correctly
gated, and assert the job-side transition through the core findById read
so the terminal half is pinned independently of the batch counters.
